+++ sdk/dotnet/Sys/SnmpTraps.cs
@@ -5,7 +5,7 @@ using System.Collections.Immutable;
 using System.Threading.Tasks;
 using Pulumi.Serialization;
 
-namespace Pulumi.F5bigip.Sys
+namespace Pulumi.F5BigIP.Sys
 {
     /// <summary>
     /// `f5bigip.sys.SnmpTraps` provides details bout how to enable snmp_traps resource on BIG-IP